+++ src/configuraptor/core.py
@@ -20,6 +20,7 @@ from .errors import (
     ConfigErrorCouldNotConvert,
     ConfigErrorInvalidType,
     ConfigErrorMissingKey,
+    FailedToLoad,
 )
 from .helpers import (
     all_annotations,
@@ -106,7 +107,9 @@ def from_url(url: str, _dummy: bool = False) -> tuple[io.BytesIO, str]:
         resp = None
         data = "{}"
     else:
-        resp = requests.get(url, timeout=10)
+        ssl_verify = os.getenv("SSL_VERIFY", "1") == "1"
+
+        resp = requests.get(url, timeout=10, verify=ssl_verify)
         data = resp.text
 
     filetype = guess_filetype_for_url(url, resp)
@@ -119,6 +122,7 @@ def _load_data(
     classname: str = None,
     lower_keys: bool = False,
     allow_types: tuple[type, ...] = (dict,),
+    strict: bool = False,
 ) -> dict[str, typing.Any]:
     """
     Tries to load the right data from a filename/path or dict, based on a manual key or a classname.
@@ -138,7 +142,9 @@ def _load_data(
 
         final_data: dict[str, typing.Any] = {}
         for source in data:
-            final_data |= load_data(source, key=key, classname=classname, lower_keys=True, allow_types=allow_types)
+            final_data |= load_data(
+                source, key=key, classname=classname, lower_keys=True, allow_types=allow_types, strict=strict
+            )
 
         return final_data
 
@@ -188,6 +194,7 @@ def load_data(
     classname: str = None,
     lower_keys: bool = False,
     allow_types: tuple[type, ...] = (dict,),
+    strict: bool = False,
 ) -> dict[str, typing.Any]:
     """
     Wrapper around __load_data that retries with key="" if anything goes wrong.
@@ -197,16 +204,18 @@ def load_data(
         data = find_pyproject_toml()
 
     try:
-        return _load_data(data, key, classname, lower_keys=lower_keys, allow_types=allow_types)
+        return _load_data(data, key, classname, lower_keys=lower_keys, allow_types=allow_types, strict=strict)
     except Exception as e:
         # sourcery skip: remove-unnecessary-else, simplify-empty-collection-comparison, swap-if-else-branches
         # @sourcery: `key != ""` is NOT the same as `not key`
         if key != "":
-            return _load_data(data, "", classname, lower_keys=lower_keys, allow_types=allow_types)
-        else:  # pragma: no cover
-            warnings.warn(f"Data could not be loaded: {e}", source=e)
-            # key already was "", just return data!
-            # (will probably not happen but fallback)
+            # try again with key ""
+            return load_data(data, "", classname, lower_keys=lower_keys, allow_types=allow_types, strict=strict)
+        elif strict:
+            raise FailedToLoad(data) from e
+        else:
+            # e.g. if settings are to be loaded via a URL that is unavailable or returns invalid json
+            warnings.warn(f"Data ('{data}') could not be loaded", source=e, category=UserWarning)
             return {}
 
 
@@ -549,7 +558,7 @@ def load_into_class(
     Shortcut for _load_data + load_into_recurse.
     """
     allow_types = (dict, bytes) if issubclass(cls, BinaryConfig) else (dict,)
-    to_load = load_data(data, key, cls.__name__, lower_keys=lower_keys, allow_types=allow_types)
+    to_load = load_data(data, key, cls.__name__, lower_keys=lower_keys, allow_types=allow_types, strict=strict)
     return _load_into_recurse(cls, to_load, init=init, strict=strict, convert_types=convert_types)
 
 
@@ -568,7 +577,7 @@ def load_into_instance(
     """
     cls = inst.__class__
     allow_types = (dict, bytes) if issubclass(cls, BinaryConfig) else (dict,)
-    to_load = load_data(data, key, cls.__name__, lower_keys=lower_keys, allow_types=allow_types)
+    to_load = load_data(data, key, cls.__name__, lower_keys=lower_keys, allow_types=allow_types, strict=strict)
     return _load_into_instance(inst, cls, to_load, init=init, strict=strict, convert_types=convert_types)